+++ benchmark-tests/benches/test_lib_bench/threads/test_lib_bench_threads.rs
@@ -12,6 +12,7 @@ use iai_callgrind::{
         .tool(Callgrind::with_args([
             "toggle-collect=*::find_primes"
         ]))
+        .tool(Dhat::default().frames(["*::find_primes"]))
 )]
 #[bench::two(2)]
 #[bench::three(3)]
@@ -24,6 +25,11 @@ fn bench_find_primes_multi_thread(num_threads: usize) -> Vec<u64> {
         .tool(Callgrind::with_args([
             "toggle-collect=thread::main",
             "toggle-collect=*::find_primes",
+        ]))
+        .tool(Dhat::default()
+            .frames([
+                "thread::main",
+                "*::find_primes"
         ]))
 )]
 #[bench::two(2)]
@@ -40,6 +46,9 @@ fn bench_thread_in_subprocess(num_threads: usize) {
         .tool(Callgrind::with_args(["--instr-atstart=no"])
             .entry_point(EntryPoint::None)
         )
+        .tool(Dhat::default()
+            .entry_point(EntryPoint::None)
+        )
 )]
 fn bench_thread_in_thread() -> Vec<u64> {
     iai_callgrind::client_requests::callgrind::start_instrumentation();
@@ -51,6 +60,9 @@ fn bench_thread_in_thread() -> Vec<u64> {
 #[library_benchmark(
     config = LibraryBenchmarkConfig::default()
         .tool(Callgrind::with_args(["instr-atstart=no"])
+            .entry_point(EntryPoint::None)
+        )
+        .tool(Dhat::default()
             .entry_point(EntryPoint::None)
         )
 )]