+++ iai-callgrind-runner/src/runner/bin_bench.rs
@@ -226,6 +226,7 @@ impl BinBench {
         let tool_configs = ToolConfigs::new(
             &mut output_format,
             config.tools,
+            &module_path,
             meta,
             default_tool,
             &EntryPoint::None,