+++ iai-callgrind-runner/tests/tests.rs
@@ -3,6 +3,8 @@ mod common;
 #[cfg(feature = "default")]
 mod test_callgrind;
 #[cfg(feature = "default")]
+mod test_dhat;
+#[cfg(feature = "default")]
 mod test_runner_binary;
 #[cfg(feature = "default")]
 mod test_tool;